fix the size of the port variable because I am the biggest idiot that has ever existed in all planes of existence

+++ include/port.h
@@ -4,6 +4,6 @@
 
 
 void outb(uint16_t port, uint8_t byte);
-uint8_t inb(uint8_t port);
+uint8_t inb(uint16_t port);
 
 #endif
